+++ src/main/backends/linux/gnome/wayland/backend.ts
@@ -35,6 +35,7 @@ export class GnomeBackend implements Backend {
    */
   public getBackendInfo() {
     return {
+      name: 'GNOME Wayland Backend',
       windowType: 'dock',
       supportsShortcuts: true,
     };